+++ src/extract/rejection.rs
@@ -41,14 +41,6 @@ define_rejection! {
 
 define_rejection! {
     #[status = BAD_REQUEST]
-    #[body = "Query string was invalid or missing"]
-    #[deprecated(since = "0.1.3", note = "No longer used. See https://github.com/tokio-rs/axum/pull/117")]
-    /// Rejection type for [`Query`](super::Query).
-    pub struct QueryStringMissing;
-}
-
-define_rejection! {
-    #[status = BAD_REQUEST]
     #[body = "Failed to parse the request body as JSON"]
     /// Rejection type for [`Json`](super::Json).
     pub struct InvalidJsonBody(BoxError);
@@ -216,7 +208,6 @@ composite_rejection! {
     /// can fail.
     pub enum QueryRejection {
         UriAlreadyExtracted,
-        QueryStringMissing,
         FailedToDeserializeQueryString,
     }
 }
@@ -228,7 +219,6 @@ composite_rejection! {
     /// can fail.
     pub enum FormRejection {
         InvalidFormContentType,
-        QueryStringMissing,
         FailedToDeserializeQueryString,
         FailedToBufferBody,
         BodyAlreadyExtracted,