+++ src/routing.rs
@@ -1,33 +1,34 @@
 //! Routing between [`Service`]s.
 
+use self::future::{BoxRouteFuture, EmptyRouterFuture, RouteFuture};
 use crate::{
     body::{box_body, BoxBody},
     buffer::MpscBuffer,
     extract::connect_info::{Connected, IntoMakeServiceWithConnectInfo},
     response::IntoResponse,
+    service::HandleErrorFromRouter,
     util::ByteStr,
 };
 use async_trait::async_trait;
 use bytes::Bytes;
-use futures_util::future;
 use http::{Method, Request, Response, StatusCode, Uri};
-use pin_project_lite::pin_project;
 use regex::Regex;
 use std::{
     borrow::Cow,
     convert::Infallible,
     fmt,
-    future::Future,
     marker::PhantomData,
-    pin::Pin,
     sync::Arc,
     task::{Context, Poll},
 };
 use tower::{
-    util::{BoxService, Oneshot, ServiceExt},
+    util::{BoxService, ServiceExt},
     BoxError, Layer, Service, ServiceBuilder,
 };
 use tower_http::map_response_body::MapResponseBodyLayer;
+
+pub mod future;
+pub mod or;
 
 /// A filter that matches one or more HTTP methods.
 #[derive(Debug, Copy, Clone)]
@@ -354,6 +355,40 @@ pub trait RoutingDsl: crate::sealed::Sealed + Sized {
     {
         IntoMakeServiceWithConnectInfo::new(self)
     }
+
+    /// Merge two routers into one.
+    ///
+    /// This is useful for breaking apps into smaller pieces and combining them
+    /// into one.
+    ///
+    /// ```
+    /// use axum::prelude::*;
+    /// #
+    /// # async fn users_list() {}
+    /// # async fn users_show() {}
+    /// # async fn teams_list() {}
+    ///
+    /// // define some routes separately
+    /// let user_routes = route("/users", get(users_list))
+    ///     .route("/users/:id", get(users_show));
+    ///
+    /// let team_routes = route("/teams", get(teams_list));
+    ///
+    /// // combine them into one
+    /// let app = user_routes.or(team_routes);
+    /// # async {
+    /// # hyper::Server::bind(&"".parse().unwrap()).serve(app.into_make_service()).await.unwrap();
+    /// # };
+    /// ```
+    fn or<S>(self, other: S) -> or::Or<Self, S>
+    where
+        S: RoutingDsl,
+    {
+        or::Or {
+            first: self,
+            second: other,
+        }
+    }
 }
 
 impl<S, F> RoutingDsl for Route<S, F> {}
@@ -381,63 +416,6 @@ where
         } else {
             let fut = self.fallback.clone().oneshot(req);
             RouteFuture::b(fut)
-        }
-    }
-}
-
-pin_project! {
-    /// The response future for [`Route`].
-    #[derive(Debug)]
-    pub struct RouteFuture<S, F, B>
-    where
-        S: Service<Request<B>>,
-        F: Service<Request<B>> {
-        #[pin] inner: RouteFutureInner<S, F, B>,
-    }
-}
-
-impl<S, F, B> RouteFuture<S, F, B>
-where
-    S: Service<Request<B>>,
-    F: Service<Request<B>>,
-{
-    pub(crate) fn a(a: Oneshot<S, Request<B>>) -> Self {
-        RouteFuture {
-            inner: RouteFutureInner::A { a },
-        }
-    }
-
-    pub(crate) fn b(b: Oneshot<F, Request<B>>) -> Self {
-        RouteFuture {
-            inner: RouteFutureInner::B { b },
-        }
-    }
-}
-
-pin_project! {
-    #[project = RouteFutureInnerProj]
-    #[derive(Debug)]
-    enum RouteFutureInner<S, F, B>
-    where
-        S: Service<Request<B>>,
-        F: Service<Request<B>>,
-    {
-        A { #[pin] a: Oneshot<S, Request<B>> },
-        B { #[pin] b: Oneshot<F, Request<B>> },
-    }
-}
-
-impl<S, F, B> Future for RouteFuture<S, F, B>
-where
-    S: Service<Request<B>, Response = Response<BoxBody>>,
-    F: Service<Request<B>, Response = Response<BoxBody>, Error = S::Error>,
-{
-    type Output = Result<Response<BoxBody>, S::Error>;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        match self.project().inner.project() {
-            RouteFutureInnerProj::A { a } => a.poll(cx),
-            RouteFutureInnerProj::B { b } => b.poll(cx),
         }
     }
 }
@@ -495,8 +473,6 @@ impl<E> Clone for EmptyRouter<E> {
     }
 }
 
-impl<E> Copy for EmptyRouter<E> {}
-
 impl<E> fmt::Debug for EmptyRouter<E> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.debug_tuple("EmptyRouter").finish()
@@ -507,7 +483,10 @@ impl<E> RoutingDsl for EmptyRouter<E> {}
 
 impl<E> crate::sealed::Sealed for EmptyRouter<E> {}
 
-impl<B, E> Service<Request<B>> for EmptyRouter<E> {
+impl<B, E> Service<Request<B>> for EmptyRouter<E>
+where
+    B: Send + Sync + 'static,
+{
     type Response = Response<BoxBody>;
     type Error = E;
     type Future = EmptyRouterFuture<E>;
@@ -516,19 +495,24 @@ impl<B, E> Service<Request<B>> for EmptyRouter<E> {
         Poll::Ready(Ok(()))
     }
 
-    fn call(&mut self, _req: Request<B>) -> Self::Future {
+    fn call(&mut self, request: Request<B>) -> Self::Future {
         let mut res = Response::new(crate::body::empty());
+        res.extensions_mut().insert(FromEmptyRouter { request });
         *res.status_mut() = self.status;
         EmptyRouterFuture {
-            future: future::ok(res),
+            future: futures_util::future::ok(res),
         }
     }
 }
 
-opaque_future! {
-    /// Response future for [`EmptyRouter`].
-    pub type EmptyRouterFuture<E> =
-        future::Ready<Result<Response<BoxBody>, E>>;
+/// Response extension used by [`EmptyRouter`] to send the request back to [`Or`] so
+/// the other service can be called.
+///
+/// Without this we would loose ownership of the request when calling the first
+/// service in [`Or`]. We also wouldn't be able to identify if the response came
+/// from [`EmptyRouter`] and therefore can be discarded in [`Or`].
+struct FromEmptyRouter<B> {
+    request: Request<B>,
 }
 
 #[derive(Debug, Clone)]
@@ -666,36 +650,6 @@ where
     }
 }
 
-pin_project! {
-    /// The response future for [`BoxRoute`].
-    pub struct BoxRouteFuture<B, E>
-    where
-        E: Into<BoxError>,
-    {
-        #[pin] inner: Oneshot<MpscBuffer<BoxService<Request<B>, Response<BoxBody>, E>, Request<B>>, Request<B>>,
-    }
-}
-
-impl<B, E> Future for BoxRouteFuture<B, E>
-where
-    E: Into<BoxError>,
-{
-    type Output = Result<Response<BoxBody>, E>;
-
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        self.project().inner.poll(cx)
-    }
-}
-
-impl<B, E> fmt::Debug for BoxRouteFuture<B, E>
-where
-    E: Into<BoxError>,
-{
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("BoxRouteFuture").finish()
-    }
-}
-
 /// A [`Service`] created from a router by applying a Tower middleware.
 ///
 /// Created with [`RoutingDsl::layer`]. See that method for more details.
@@ -812,7 +766,7 @@ impl<S> Layered<S> {
     pub fn handle_error<F, ReqBody, ResBody, Res, E>(
         self,
         f: F,
-    ) -> crate::service::HandleError<S, F, ReqBody>
+    ) -> crate::service::HandleError<S, F, ReqBody, HandleErrorFromRouter>
     where
         S: Service<Request<ReqBody>, Response = Response<ResBody>> + Clone,
         F: FnOnce(S::Error) -> Result<Res, E>,